feat: build html and css structure
add arrow and circle selector navigation buttons
and sample images with an active, centered image
to template.html.

style the html in style.css by making the active
carousel image larger and more prominent, and
have the active button selector be filled to
indicated which image you are on.

update README.md.