feat: build out JS, clean up html
remove placeholder html data, now generated via
the DOM.

minor change to style.css to fix arrow positioning
in relation to the carousel.

add the following carousel functionality code to
index.js:
add image array and set the number of images
visible on the screen at any given time.
set the the center-most carousel image as the
starting point.
create function renderCarousel() that builds the
carousel images and sets the starting image as the
'active' featured image.
create function renderSelectors() that builds the
selector circles for each image in imageSources
and sets the 'active' circle based on the current
centerIndex image, also listen for clicks on
circles and then sets centerIndex to that circle's
index and rerenders carousel and the selectors.

create functions nextImage() and prevImage() that
set centerIndex to either one more or less and
rerender the carousel.

add eventListeners to both left and right arrows
to run the next/prevImage() functions.